+++ lib/spectrum/json.rb
@@ -87,8 +87,8 @@ require 'spectrum/request/debug'
 
 require 'spectrum/policy/get_this'
 
-require 'spectrum/presenters/holdings/holding_presenter'
-require 'spectrum/presenters/holdings/mirlyn_item_presenter'
+require 'spectrum/presenters/holding_presenter'
+require 'spectrum/presenters/mirlyn_item_presenter'
 require 'spectrum/holding/mirlyn_item_description'
 require 'spectrum/holding/action'
 require 'spectrum/holding/get_this_action'
