+++ lib/spectrum/presenters/holding_presenter.rb
