+++ lib/spectrum/presenters/mirlyn_item_presenter.rb
